Handle close message by exiting the transport
Before this change, the client would end up
receiving an empty message when the close message
was received. This isn't good because the client
expects JSON and an empty string would cause an
error.

This change adds a check for the close message
and terminates the transport when it is received.
This allows the client to retry if needed.

This also has a nice side effect of resolving an
additional error scenario where we weren't properly
handling the "unwrap" of the message that has been
factored out with this commit.

This change also adds some debug output to help
understand the types of messages that are being
received. It will not be included in the release
build.

The code could definitely be cleaned up a bit, but
that's a task for another day once we've gotten a
bit further along.

Co-authored-by: Nate Vick <nate.vick@hint.io>